Revert "FF refactor for internal teams (#1490)" (#1493)
This reverts commit ec004bff55b52b0829cd83a08ea55305cdfc1ab7.
