--- web/scenes/Portal/Teams/TeamId/Apps/AppId/Configuration/layout/constants-temp.ts
@@ -1,8 +1,0 @@
-/** show refactored forms to these team ids for beta tests */
-export const INTERNAL_TEAM_IDS = [
-  "team_4e67539b4bb0f6dfabeba48793cf747d", //TFH
-  "team_4e67539b4bb0f6dfabeba48793cf747d", //Michal's
-  "team_1767d2864edd0a422e0974f4a8a406e3", // staging TFH
-  "team_b250185634d599fbea3f151379ddcd19", // staging Michal's
-  "team_09a9c2ea77728bd317a5f4c5f2c659c2", // local
-];
